requirements.txt
cpython-lldb==0.3.2
openai==0.28.0
psutil==5.9.5
PyYAML==6.0.1
selenium==4.9.0